avoid integer overflow in container size precheck
the container size precheck multiplied a wire-controlled element count by
the per-element minimum size in 32-bit arithmetic, so a crafted list, set
or map header could wrap the product down to a small or zero value and slip
past the maxMessageSize guard before any allocation was bounded.

widen the multiplication to a fixed-width 64-bit type and widen the
checkReadBytesAvailable parameter to match, in both the c++ protocols and
the c_glib binary and compact readers, so the product can no longer overflow
on LLP64 platforms such as Windows where long is only 32-bit. regression
tests cover the three c++ protocols.

Client: cpp,c_glib

+++ lib/c_glib/src/thrift/c_glib/protocol/thrift_binary_protocol.c
@@ -577,9 +577,9 @@ thrift_binary_protocol_read_map_begin (ThriftProtocol *protocol,
     return -1;
   }
 
-  if(!ttc->checkReadBytesAvailable (THRIFT_TRANSPORT(tp->transport), 
-                                    sizei * thrift_binary_protocol_get_min_serialized_size(protocol, k, error) + 
-                                    sizei * thrift_binary_protocol_get_min_serialized_size(protocol, v, error),
+  if(!ttc->checkReadBytesAvailable (THRIFT_TRANSPORT(tp->transport),
+                                    (gint64) sizei * thrift_binary_protocol_get_min_serialized_size(protocol, k, error) +
+                                    (gint64) sizei * thrift_binary_protocol_get_min_serialized_size(protocol, v, error),
                                     error))
   {
     return -1;
@@ -642,8 +642,8 @@ thrift_binary_protocol_read_list_begin (ThriftProtocol *protocol,
     return -1;
   }
 
-  if(!ttc->checkReadBytesAvailable (THRIFT_TRANSPORT(tp->transport), 
-                                    (sizei * thrift_binary_protocol_get_min_serialized_size(protocol, e, error)),
+  if(!ttc->checkReadBytesAvailable (THRIFT_TRANSPORT(tp->transport),
+                                    ((gint64) sizei * thrift_binary_protocol_get_min_serialized_size(protocol, e, error)),
                                     error))
   {
     return -1;
